completed chapter 5
