+++ package.json
@@ -14,5 +14,8 @@
   "type": "module",
   "workspaces": [
     "src/icp_frontend"
-  ]
+  ],
+  "dependencies": {
+    "react-router-dom": "^6.23.1"
+  }
 }
